Conflicts:
	tasks.py

+++ tasks.py
@@ -1,22 +1,27 @@
-import os
+import base64
 import datetime
 import djcelery
-import pickle
-import urllib2
-import base64
-import urllib
-import redis
 import ewrl
-from django.conf import settings
-from celery.task.control import inspect
+import os
+import pickle
+import urllib
+import urllib2
 from celery.decorators import task
-from celery.signals import worker_ready,beat_init
 from celery.execute import send_task
-from kral.views import push_data
+from celery.signals import worker_ready,beat_init
+from celery.task.control import inspect
+from django.conf import settings
 from eventlet.timeout import Timeout
 from models import Feed
+from kral.views import push_data
 
-cache = redis.Redis(host='localhost', port=6379, db=1)
+try:
+    import redis
+    cache = redis.Redis(host='localhost', port=6379, db=1)
+except ImportError:
+    redis = False
+    from django.core.cache import cache
+
 
 ALLPLUGINS = []
 if not hasattr(settings, "KRAL_PLUGINS"): 
@@ -35,8 +40,11 @@ def kral_init(**kwargs):
     from djcelery.models import PeriodicTask,PeriodicTasks
     PeriodicTask.objects.all().delete()
     PeriodicTasks.objects.all().delete()
-    task_cache = redis.Redis(host='localhost', port=6379, db=0)
-    task_cache.flushdb()
+    if redis:
+        task_cache = redis.Redis(host='localhost', port=6379, db=0)
+        task_cache.flushdb()
+    else:
+        task_cache = cache
     cache.set('KRAL_QUERIES',settings.KRAL_QUERIES)
 beat_init.connect(kral_init)
